+++ main.js
@@ -16,6 +16,8 @@ import { downloadStopTimes, fetchStopTimes } from "./src/api/stopTimes.js";
 import { downloadStops, fetchStops } from "./src/api/stops.js";
 import { downloadTrips, fetchTrips } from "./src/api/trips.js";
 const statusDownload = document.getElementById("status-download");
+const statusZip = document.getElementById("status-zip");
+const saveUi = document.getElementById("save-ui");
 
 const checkApiKey = document.getElementById("check-api-key");
 const downloadFilesUi = document.getElementById("download-files-ui");
@@ -23,6 +25,36 @@ checkApiKey.addEventListener("click", async function () {
   selectAgency.innerHTML = "<option>Choose an agency</option>";
   await fetchAgency("agency");
 });
+
+async function createZip() {
+  const zip = new JSZip();
+  const folder = zip.folder("zeusdata");
+
+  let vehicles = await fetchVehicles("vehicles");
+  let routes = await fetchRoutes("routes");
+  let trips = await fetchTrips("trips");
+  let shapes = await fetchShapes("shapes");
+  let stops = await fetchStops("stops");
+  let stopTimes = await fetchStopTimes("stop_times");
+
+  folder.file("vehicles.json", JSON.stringify(vehicles, null, 2));
+  folder.file("routes.json", JSON.stringify(routes, null, 2));
+  folder.file("trips.json", JSON.stringify(trips, null, 2));
+  folder.file("shapes.json", JSON.stringify(shapes, null, 2));
+  folder.file("stops.json", JSON.stringify(stops, null, 2));
+  folder.file("stopTimes.json", JSON.stringify(stopTimes, null, 2));
+
+  const content = await zip.generateAsync({ type: "blob" });
+  // let blob = new Blob([vehicles], { type: "application/json" });
+  return content;
+  // const a = document.createElement("a");
+  //
+  // a.href = url;
+  // a.download = "zeus_data_tranzty_files.zip";
+  // a.click();
+  //
+  // URL.revokeObjectURL(url);
+}
 
 const generateData = document.getElementById("generate-data");
 
@@ -32,6 +64,7 @@ generateData.addEventListener("click", async function () {
     statusData.textContent = "ERROR: AGENCY REQUIRED!";
     selectUi.style.border = "1px solid crimson";
   } else {
+    createZip();
     await fetchVehicles("vehicles");
     await fetchRoutes("routes");
     await fetchTrips("trips");
@@ -41,9 +74,12 @@ generateData.addEventListener("click", async function () {
     selectUi.style.border = "1px solid #67bf82";
     statusData.textContent = "DATA GENERATED!";
     statusDownload.textContent = `FILES FOR AGENCY ID: ${selectedAgency}`;
+    statusZip.textContent = `ZIP FILE FOR AGENCY ID: ${selectedAgency}`;
     setInterval(() => {
       downloadFilesUi.style.transition = "1s ease-in-out";
       downloadFilesUi.style.border = "1px solid #67bf82";
+      saveUi.style.transition = "1s ease-in-out";
+      saveUi.style.border = "1px solid #67bf82";
     }, 100);
   }
 });
@@ -77,5 +113,47 @@ selectAgency.addEventListener("click", function () {
   ) {
     selectUi.style.border = "1px solid crimson";
     statusData.textContent = "KEY NEEDED!";
+  }
+});
+//
+const save = document.getElementById("btn-save");
+// save.addEventListener("click", function () {
+//   const zip = new JSZip();
+//   const folder = zip.folder("zeusdata");
+//
+//   folder.file("readme", "text");
+//   zip.generateAsync({ type: "blob" }).then(function (content) {
+//     const url = URL.createObjectURL(content);
+//
+//     const a = document.createElement("a");
+//     a.href = url;
+//     a.download = "name.zip";
+//     a.click();
+//
+//     URL.revokeObjectURL(url);
+//   });
+// });
+save.addEventListener("click", async function () {
+  if (
+    (selectedAgency === null && statusKey.textContent == "") ||
+    statusKey.textContent == "ERROR:INVALID KEY!"
+  ) {
+    statusZip.textContent = "ERROR: FILES NOT GENERATED!";
+    apiKeyUi.style.border = "1px solid crimson";
+    saveUi.style.border = "1px solid crimson";
+    selectUi.style.border = "1px solid crimson";
+    downloadFilesUi.style.border = "1px solid crimson";
+    saveUi.style.border = "1px solid crimson";
+  } else {
+    const blob = await createZip();
+
+    const url = URL.createObjectURL(blob);
+    const a = document.createElement("a");
+
+    a.href = url;
+    a.download = "zeus_data_tranzty_files.zip";
+    a.click();
+
+    setTimeout(() => URL.revokeObjectURL(url), 100);
   }
 });