+++ CHANGELOG.md
@@ -14,3 +14,7 @@
 ### Improved
 - UI/UX refinements
 - Better user flow and feedback
+
+## v0.2.0 - JSZip (29-04-2026)
+- FEATURE: Integrate JSZip library to export fetched data as .zip file
+- FIX: Solved an issue where download was triggered after selecting an agency